+++ frontend/src/lib/components/food-scan/FoodScan.svelte
@@ -1,21 +1,26 @@
 <script lang="ts">
   import { executeOcrProcess } from '$lib/api/ocr.api';
+  import { themeStore } from '$lib/util/themeStore';
   import { initialized, t } from '$lib/util/translations';
   import * as webllm from '@mlc-ai/web-llm';
   import Card from '@smui/card';
   import Checkbox from '@smui/checkbox';
+  import Chip, { Set, Text } from '@smui/chips';
   import FormField from '@smui/form-field';
-  import Select, { Option } from '@smui/select';
+  import ContentOutput from './ContentOutput.svelte';
+  import DebugInformation from './DebugInformation.svelte';
 
   const { prebuiltAppConfig } = webllm;
   const { model_list: availableModels } = prebuiltAppConfig;
+  // console.log(availableModels.map((model, i) => `${i} ${model.model_id}`));
   const defaultSystemPrompt = `Du erhältst einen Text der auf der Rückseite der Verpackung eines Nahrungsmittels steht. Du antwortest darauf ausschließlich mit der Bezeichnung des Lebensmittels und ob es vegan ist (ja / nein)! Trenne die beiden Informationen durch ein Semikolon.`;
+  const followUpQuestionSystemPrompt = `Du erhältst eine Frage zum Lebensmittel, das du vorhin analysiert hast. Du antwortest darauf ausschließlich mit einer klaren und prägnanten Antwort.`;
 
   let engine: webllm.MLCEngineInterface;
-  let selectedModel = availableModels[52]; // Qwen2.5-3B-Instruct-q4f32_1-MLC
+  let selectedModel = availableModels[4]; // Llama-3.2-3B-Instruct-q4f32_1-MLC
   let systemPromptText = defaultSystemPrompt;
   let userPromptText = '';
-  let llmResults = [];
+  let llmResult = '';
   let promptResStats: any = {};
   let inputFiles: File[] = [];
   let imagePreviewSrc = '';
@@ -38,6 +43,16 @@
     generateUserPrompt();
   }
   $: if (userPromptText) promptLLM();
+  $: diet =
+    llmResult?.split(';')?.[1]?.toLowerCase() === 'ja'
+      ? 'vegan'
+      : 'nicht vegan';
+  $: followUpQuestions = [
+    `Warum ist dieses Lebensmittel ${diet}?`,
+    'Ist das Lebensmittel gesund?',
+    'Ist das Lebensmittel laktosefrei?',
+    'Ist das Lebensmittel glutenfrei?',
+  ];
 
   const modelInitProgressCallback = (report: webllm.InitProgressReport) => {
     loadModelInfo = report.text;
@@ -74,15 +89,27 @@
     userPromptText = ocrResponse?.ParsedResults?.[0]?.ParsedText;
   }
 
-  async function promptLLM() {
+  async function promptLLM(followUpQuestion: string = '') {
     console.log('Prompting LLM...');
     const promptStartTime = Date.now();
+    let messages: any = [];
 
-    const promptResponse = await engine.chat.completions.create({
-      messages: [
+    if (followUpQuestion) {
+      messages = [
+        { role: 'system', content: followUpQuestionSystemPrompt },
+        { role: 'user', content: userPromptText },
+        { role: 'assistant', content: llmResult },
+        { role: 'user', content: followUpQuestion },
+      ];
+    } else {
+      messages = [
         { role: 'system', content: systemPromptText },
         { role: 'user', content: userPromptText },
-      ],
+      ];
+    }
+
+    const promptResponse = await engine.chat.completions.create({
+      messages,
       // below configurations are all optional
       n: 1,
       temperature: 0.01,
@@ -104,7 +131,12 @@
     console.log(`LLM responded within ${llmResponseTime} seconds.`, {
       promptResponse,
     });
-    llmResults = promptResponse.choices;
+
+    if (followUpQuestion) {
+      llmResult += `;${promptResponse.choices[0].message.content}`;
+    } else {
+      llmResult = promptResponse.choices[0].message.content;
+    }
     promptResStats = promptResponse.usage;
 
     loading = false;
@@ -129,13 +161,24 @@
     }
   }
 
+  async function askFollowUpQuestion(question: string) {
+    loading = true;
+    promptLLM(question);
+  }
+
   main();
 </script>
 
 {#if $initialized}
   <section>
     {#if !engineReady}
-      <Card padded class="warning_hint">
+      <Card
+        padded
+        class="warning_hint"
+        style={$themeStore === 'dark'
+          ? 'background-color: var(--color_bg_2)'
+          : 'background-color: var(--color_bg_light_2)'}
+      >
         <h2>{$t('page.foodScan.loadingHint')}</h2>
         <p class="warning_sign">⚠️</p>
         <p>{@html $t('page.foodScan.warning')}</p>
@@ -143,100 +186,84 @@
       </Card>
     {:else}
       <h3>{$t('page.foodScan.introduction')}</h3>
-      <label for="file-upload" class="custom_file_label"
-        >{$t('page.foodScan.chooseFile')}</label
-      >
-      <input
-        id="file-upload"
-        class="file_input"
-        type="file"
-        accept="image/png, image/jpeg"
-        bind:files={inputFiles}
-      />
+      <div class="file_inputs">
+        <form>
+          <label
+            for="file_upload"
+            class="custom_file_label"
+            style={$themeStore === 'dark'
+              ? 'background-color: var(--color_bg_1)'
+              : 'background-color: var(--color_bg_light_1)'}
+            >{$t('page.foodScan.chooseFile')}</label
+          >
+          <input
+            id="file_upload"
+            class="file_input"
+            type="file"
+            accept="image/*"
+            bind:files={inputFiles}
+          />
+        </form>
+
+        <p>{$t('page.foodScan.or')}</p>
+
+        <form>
+          <label for="camera_upload" class="custom_file_label"
+            >{$t('page.foodScan.takePicture')}</label
+          >
+          <input
+            id="camera_upload"
+            class="file_input"
+            type="file"
+            accept="image/*"
+            bind:files={inputFiles}
+            capture
+          />
+        </form>
+      </div>
 
       {#if inputFiles?.length > 0}
-        <div class="initial_scan_output">
-          <img
-            class="image_preview"
-            src={imagePreviewSrc || ''}
-            alt="Preview"
-          />
+        <ContentOutput
+          {imagePreviewSrc}
+          {loading}
+          {inputFiles}
+          {llmResult}
+          {diet}
+        />
 
-          <div class="prompt_output">
-            {#if loading && inputFiles.length != 0}
-              <p>{$t('page.foodScan.loading')}</p>
-            {:else if !loading && llmResults.length != 0}
-              {#each llmResults as responseTexts, index (index)}
-                <h2>{responseTexts.message.content.split(';')[0]}</h2>
-                <p id="generate-label">
-                  {#if responseTexts.message.content
-                    .split(';')[1]
-                    .toLowerCase()
-                    .includes('ja')}
-                    {$t('page.foodScan.vegan')}
-                  {:else}
-                    {$t('page.foodScan.notVegan')}
-                  {/if}
-                </p>
-              {/each}
-            {/if}
-          </div>
-        </div>
+        {#if llmResult}
+          <Set chips={followUpQuestions} let:chip>
+            <Chip
+              {chip}
+              style={$themeStore === 'dark'
+                ? 'background-color: var(--color_bg_1)'
+                : 'background-color: var(--color_bg_light_1)'}
+              on:click={() => askFollowUpQuestion(chip)}
+              ><Text>{chip}</Text></Chip
+            >
+          </Set>
+        {/if}
+
         <FormField>
           <Checkbox bind:checked={debugInfoActive} />
           <p slot="label">{$t('page.foodScan.technicalInfo')}</p>
         </FormField>
 
         {#if debugInfoActive}
-          <div>
-            <h3>{$t('page.foodScan.debug.title')}</h3>
-            <p>
-              {$t('page.foodScan.debug.selectedModel', {
-                modelName: selectedModel?.model_id,
-              })}
-            </p>
-            <Select
-              class="select_model"
-              bind:value={selectedModel}
-              label="Select Model"
-              on:SMUISelect:change={reloadModel}
-            >
-              {#each availableModels as model}
-                <Option value={model}
-                  >{model?.model_id} ({model.vram_required_MB}MB VRAM / {model
-                    ?.overrides?.context_window_size} window size)</Option
-                >
-              {/each}
-            </Select>
-
-            <p>{$t('page.foodScan.debug.systemPrompt')}</p>
-            <Card padded>{systemPromptText}</Card>
-
-            <p>{$t('page.foodScan.debug.userPrompt')}</p>
-            <Card padded>{userPromptText}</Card>
-
-            <p>{$t('page.foodScan.debug.stats.title')}</p>
-            <table>
-              <tr>
-                <th>{$t('page.foodScan.debug.stats.ocr')}</th>
-                <th>{$t('page.foodScan.debug.stats.llm')}</th>
-                <th>{$t('page.foodScan.debug.stats.completion')}</th>
-                <th>{$t('page.foodScan.debug.stats.prompt')}</th>
-                <th>{$t('page.foodScan.debug.stats.total')}</th>
-                <th>{$t('page.foodScan.debug.stats.prefill')}</th>
-                <th>{$t('page.foodScan.debug.stats.decode')}</th>
-              </tr>
-              <tr>
-                <td>{ocrResponseTimeText}</td>
-                <td>{llmResponseTimeText}</td>
-                <td>{completionTokens}</td>
-                <td>{promptTokens}</td>
-                <td>{totalTokens}</td>
-                <td>{prefillTokensPerS}</td>
-                <td>{decodeTokensPerS}</td>
-              </tr>
-            </table>
-          </div>
+          <DebugInformation
+            {selectedModel}
+            {availableModels}
+            {systemPromptText}
+            {userPromptText}
+            {ocrResponseTimeText}
+            {llmResponseTimeText}
+            {completionTokens}
+            {promptTokens}
+            {totalTokens}
+            {prefillTokensPerS}
+            {decodeTokensPerS}
+            on:reloadModel={reloadModel}
+          />
         {/if}
       {/if}
     {/if}
@@ -246,6 +273,8 @@
 {/if}
 
 <style lang="scss">
+  @import '../../styles/variables.scss';
+
   section {
     padding-top: 2rem;
     display: flex;
@@ -268,6 +297,17 @@
     margin: 0;
   }
 
+  .file_inputs {
+    display: flex;
+    gap: 2rem;
+    justify-content: center;
+    align-items: center;
+
+    @media #{$phone} {
+      gap: 1rem;
+    }
+  }
+
   .file_input {
     opacity: 0;
     position: absolute;
@@ -277,65 +317,11 @@
   .custom_file_label {
     display: inline-block;
     padding: 10px 15px;
-    background-color: var(--black30);
-    color: white;
     border-radius: 5px;
     cursor: pointer;
   }
 
-  .initial_scan_output {
-    display: flex;
-    flex-direction: row;
-    gap: 2rem;
-    margin-top: 1rem;
-  }
-
-  .prompt_output {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding: 0.25rem 0.25rem 0 0;
-
-    p {
-      text-align: left;
-    }
-  }
-
-  h3 {
-    margin: 0;
-  }
-
-  .image_preview {
-    max-width: 33vw;
-    box-shadow:
-      0 4px 8px rgba(0, 0, 0, 0.2),
-      0 6px 20px rgba(0, 0, 0, 0.19);
-    border-radius: 10px;
-    border: 3px solid var(--black30);
-    transition:
-      transform 0.5s ease,
-      box-shadow 0.5s ease;
-
-    &:hover {
-      transform: scale(1.05);
-      box-shadow: 0 8px 16px rgba(0, 0, 0, 0.3);
-    }
-  }
-
-  :global(.select_model) {
-    width: 50vw;
-  }
-
-  table {
-    font-family: arial, sans-serif;
-    border-collapse: collapse;
-    width: 100%;
-  }
-
-  td,
-  th {
-    border: 1px solid #dddddd;
-    text-align: left;
-    padding: 8px;
+  button {
+    all: unset;
   }
 </style>