+++ frontend/src/routes/food-scan/+page.svelte
@@ -2,10 +2,17 @@
   import FoodScan from '$lib/components/food-scan/FoodScan.svelte';
   import ToggledApplicationInfo from '$lib/components/shared/ToggledApplicationInfo.svelte';
   import { applicationRoutes } from '$lib/config/applications';
-  import { getlocale } from '$lib/util/translations';
+  import { languageStore } from '$lib/util/languageStore';
+  import { setLocale } from '$lib/util/translations';
+  import { onMount } from 'svelte';
 
   const { 'food-scan': foodScanRoute } = applicationRoutes;
-  const locale = getlocale();
+
+  $: locale = $languageStore;
+
+  onMount(async () => {
+    await setLocale($languageStore);
+  });
 </script>
 
 <svelte:head>